requirements.txt
attrs==24.2.0
canmatrix==1.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
future==1.0.0
kiwisolver==1.4.5
matplotlib==3.9.0
msgpack==1.0.8
numpy==1.26.4
obd==0.7.2
packaging==24.1
pandas==2.2.2
pillow==10.3.0
Pint==0.20.1
pyDatalog==0.17.4
pyparsing==3.1.2
pyserial==3.5
python-can==4.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyVISA==1.13.0
PyVISA-py==0.7.0
seaborn==0.13.2
six==1.16.0
typing_extensions==4.7.1
tzdata==2024.1
websocket-client==1.8.0
wrapt==1.16.0
